Change path manipulation to use Path
NOTE: This might have changed behavior in several places, like exisiting
usage of `AsFileName(..)` which has a breaking change. See below.

Using string concatenation to work on path easily leads to bugs with
missing or duplicate path separators.
- "p" + "f.ext" is "pf.ext" rather than "p/f.ext"
- "p/" + "/f.ext" is "p//f.ext" rather than "p/f.ext"

Path.Combine handles such cases, and additionally the problem where \ is
used as a path separator as is often the case on Windows.

This potentionally buggy path handling was found when setting up Altinn
Studio the first time, setting a path without trailing / where a
trailing / was expected.

While the changes should be quite safe in theory, there might be
existing assumptions which makes it fail. One such is the assumption
that / will be the directory separator in files, which is not the case
after combining paths on operating systems which uses a different
primary path separator (like Windows).

There is additionally a breaking change in `AsFileName()` where Path
separators are replaced with _ rather than -.